move PersonBanner in person_view.dart as a private class
